projects/weather/client/requirements.txt
requests
tomorrow_io
time
paho-mqtt